[Refactor] Externalモジュール内に移動
