+++ app/forms/tweet_form.rb
@@ -10,7 +10,7 @@ class TweetForm
     validate!
 
     if Rails.env.production?
-      TwitterClient.new.post("#{text}\n#{url}")
+      External::TwitterClient.new.post("#{text}\n#{url}")
     else
       Rails.logger.info "Post to twitter.\n#{text}"
     end
