+++ spec/models/external/mattermost_client_spec.rb
@@ -1,13 +1,13 @@
 require 'rails_helper'
 
-RSpec.describe MattermostClient do
+RSpec.describe External::MattermostClient do
   describe '.post' do
     it 'Faraday::Connection のインスタンスにpostしていること' do
       client_double = instance_double(Faraday::Connection)
       allow(Faraday::Connection).to receive(:new).and_return(client_double)
       expect(client_double).to receive(:post)
 
-      MattermostClient.post({ payload: 'payload' })
+      External::MattermostClient.post({ payload: 'payload' })
     end
   end
 end
